+++ source/TestCommon/source/TestCommon.Tests/Unit/AutoFixture/AttributeTests/InlineAutoMoqDataAttributeTests.cs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
 using Energinet.DataHub.Core.TestCommon.AutoFixture.Attributes;
 using FluentAssertions;
 using Xunit;